+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   in their name.
 - Quirks mode: sort matches by reversed document path (document first)
 - Node names/paths where double encoded both when importing them and when executing the "find" function
-- Quirks mode: emulate C locale collation for node name
+- Quirks mode: use default collation of Rust for corpora imported from relANNIS 3.3
 
 ## [0.20.0] - 2019-05-19